Added two queries and tested for database management, as well as refactoring connection and schema.
Removed not null from order confirmed and completion times.
Added email address to users.

Refactored database connection to be more modular, moved into its own file that generates an object with a query function, and passed dbParams to it on require.
Created queries for all orders for a restaurant, as well as all items on an order.